authenticatioin and authorization

+++ controllers/api/v1/posts_api.js
@@ -23,20 +23,21 @@ module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
 
-    // if (post.user == req.user.id) {
-    post.remove();
+    if (post.user == req.user.id) {
+      post.remove();
 
-    await Comment.deleteMany({
-      post: req.params.id,
-    });
+      await Comment.deleteMany({
+        post: req.params.id,
+      });
 
-    return res.status(200).json({
-      message: "Posts and Associated Comments Deleted Successfully!",
-    });
-    // } else {
-    //   req.flash("error", "You cannot delete this post");
-    //   return res.redirect("back");
-    // }
+      return res.status(200).json({
+        message: "Posts and Associated Comments Deleted Successfully!",
+      });
+    } else {
+      return res.status(401).json({
+        message: "You can't delete this post!",
+      });
+    }
   } catch (err) {
     console.log("*****", err);
     return res.status(500).json({
